Rename Did -> ManagedDID, ParsedDid -> DID
The terminology within the SDK is getting confusing. I want to standardized a little bit.

A DID is now a struct that you can grab various pieces of information from:
* uri
* methodName
* methodSpecificId
* etc.

A ManagedDID will also be a DID, but it has its keys managed by a .
