+++ Sources/Web5/Dids/ManagedDID.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol ManagedDID {
+    var uri: String { get }
+    var keyManager: any KeyManager { get }
+}
