+++ Tests/Web5Tests/Dids/ParsedDidTests.swift
@@ -6,7 +6,7 @@ class ParsedDidTests: XCTestCase {
 
     func test_initValidUri() throws {
         let didUri = "did:example:123abc"
-        let parsed = try ParsedDid(didUri: didUri)
+        let parsed = try DID(didUri: didUri)
         XCTAssertEqual(parsed.uri, didUri)
         XCTAssertEqual(parsed.methodName, "example")
         XCTAssertEqual(parsed.methodSpecificId, "123abc")
@@ -14,7 +14,7 @@ class ParsedDidTests: XCTestCase {
 
     func test_initWithDidWebUriThatContainsPath() throws {
         let didUri = "did:web:w3c-ccg.github.io:user:alice"
-        let parsed = try ParsedDid(didUri: didUri)
+        let parsed = try DID(didUri: didUri)
         XCTAssertEqual(parsed.uri, didUri)
         XCTAssertEqual(parsed.methodName, "web")
         XCTAssertEqual(parsed.methodSpecificId, "w3c-ccg.github.io:user:alice")
@@ -22,6 +22,6 @@ class ParsedDidTests: XCTestCase {
 
     func test_initInvalidUri() throws {
         let didUri = "invalid:uri"
-        XCTAssertThrowsError(try ParsedDid(didUri: didUri))
+        XCTAssertThrowsError(try DID(didUri: didUri))
     }
 }
